+++ OUDS/Core/OUDS/Sources/OUDSTheme/OUDSTheme+SemanticTokens/OUDSTheme+OpacitySemanticTokens.swift
@@ -23,6 +23,6 @@ extension OUDSTheme: OpacitySemanticTokens {
     @objc open var opacityWeaker: OpacitySemanticToken { OpacityRawTokens.opacity100 }
     @objc open var opacityWeak: OpacitySemanticToken { OpacityRawTokens.opacity300 }
     @objc open var opacityMedium: OpacitySemanticToken { OpacityRawTokens.opacity500 }
-    @objc open var opacityEmphasized: OpacitySemanticToken { OpacityRawTokens.opacity700 }
+    @objc open var opacityStrong: OpacitySemanticToken { OpacityRawTokens.opacity700 }
     @objc open var opacityOpaque: OpacitySemanticToken { OpacityRawTokens.opacity900 }
 }