+++ OUDS/Core/OUDS/Tests/OUDSTheme/MockTheme/MockTheme+OpacitySemanticTokens.swift
@@ -23,6 +23,6 @@ extension MockTheme {
     override var opacityWeaker: OpacitySemanticToken { Self.mockThemeOpacityRawToken }
     override var opacityWeak: OpacitySemanticToken { Self.mockThemeOpacityRawToken }
     override var opacityMedium: OpacitySemanticToken { Self.mockThemeOpacityRawToken }
-    override var opacityEmphasized: OpacitySemanticToken { Self.mockThemeOpacityRawToken }
+    override var opacityStrong: OpacitySemanticToken { Self.mockThemeOpacityRawToken }
     override var opacityOpaque: OpacitySemanticToken { Self.mockThemeOpacityRawToken }
 }